Move default settings into plugin
